requirements.txt
asttokens==3.0.0
beautifulsoup4==4.14.2
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
flake8==7.3.0
fsspec==2025.12.0
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
Markdown==3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.6
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyMuPDF==1.26.6
pypdf==6.2.0
pytest==9.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
faiss-cpu==1.9.0
